requirements.txt
boto3
requests
pandas
snowflake-connector-python
streamlit
plotly